Correct conversion service naming converntions
* package is now org.scijava.convert. Shouldn't have been a subpackage
  of *.util.
* ConversionHandler is now just a Converter, to imply action.
* ConversionService is now just ConvertService.
